pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "safe-trading-agent-template"
version = "0.1.0"
description = "LangGraph starter for autonomous trading agents gated by Headless Oracle market verification"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["langgraph", "trading-agent", "headless-oracle", "market-safety", "ed25519", "langchain"]

dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.2",
    "langchain-anthropic>=0.3",
    "httpx>=0.27",
    "PyNaCl>=1.5",
    "python-dotenv>=1.0",
    # Headless Oracle Python SDK — remove after PyPI publish and replace with headless-oracle>=0.1.0
    "headless-oracle @ git+https://github.com/LembaGang/headless-oracle-python.git",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-mock>=3.12",
]

[project.urls]
Homepage    = "https://headlessoracle.com"
Repository  = "https://github.com/LembaGang/safe-trading-agent-template"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: tests that make live network calls to headlessoracle.com (deselect with -m 'not integration')",
]
